requirements.txt
pexpect
python-vagrant
paramiko